Abstract firebase from controllers
This pushes the firebase logic behind services and keeps it out of the controllers. This is because, long term, I'd like to move away from using firebase and implement my own backend. Also, this strips out all the (admittedly trivial) unit tests. I'm severly missing access to the sweet testing library we have at work. Combine that with our short timeline for initial release, and I'm willing to let the tests slide for now until we have time to come back and do them right without a ton of repeated boilerplate code.