Remove concept of multiple clubs
Let's start simple and only have a single club. I've replaced the clubs and club pages with a single dashboard that shows what was on the old club page. I've removed the concept of multiple clubs from the app and Firebase. I've added a properties object that will hold things we might want to change on the fly later, for now this is just the club name.